db_api: replace generic get_one() with more specific get_deposit(), ...
This makes the api a little more intuitive and the using modules
of the api do not need to know the tablenames anymore. The table
names should be hidden in db_api to keep the scope of the variables
small as possible. This could later make table name changes a lot
easier.