db_api: replace generic get_all() with more specific list_deposits(), ...
This makes the api a little more intuitive and the external modules
using the api do not need to know the table names anymore. The table
names should be hidden in db_api to keep the scope of the variables
as small as possible. This will make changing the table names a lot
easier.