Introducing the inventory system.
Products now have the property "countable".
If countable is true, the stock field increases or decreases every purchase.

PLEASE NOTICE:
There is no database migration script yet! Do not pull this commit to any
productive system, it won't work! The database scheme must be updated
beforehand.
Furthermore, the frontend and the admin panel don't have an update, either.
This must be done quickly. I'll open an issue for this in both projects.

fixes #27