Introducing departmentpurchases
Departments can now register their own purchases.
The amount of these purchases is charged to them as payoff
and the stock of the purchased products is updated.

Payoffs now have an additional field "Departmentpurchase_id",
so that in case of a payoff return the product stock is updated.

Several minor changes, Linter,....