"Departmentpurchases" are now stored in "DepartmentpurchaseCollections".
Each individual purchase has only the number of products purchased and
the *total price* of these products.
This fixes the bug that products purchased together in a value pack lead to
rounding errors when calculating the unit price.
Each "DepartmentpurchaseCollection" contains one or more "Departmentpurchases".

*IMPORTANT*
It may be necessary for previous departmentpurchases to be deleted. This must be done
by hand. The update script "./project/updates/0_departmentpurchasecollections.py" can be
used for this purpose. Copy this file and the corresponding SQL file to the main
shop.db directory and execute it.